requirements.txt
tqdm>=4.66.5,<5.0.0
scipy>=1.14.1,<2.0.0
numpy>=1.26.4,<2.0.0
uvicorn>=0.30.5,<1.0.0
fasttext>=0.9.3,<1.0.0
pydantic>=2.9.0,<3.0.0
fastapi>=0.114.0,<1.0.0
requests>=2.32.3,<3.0.0
python-dotenv>=1.0.1,<2.0.0